+++ src/roiextractors/extractors/caiman/caimansegmentationextractor.py
@@ -5,6 +5,7 @@ Classes
 CaimanSegmentationExtractor
     A class for extracting segmentation from CaImAn output.
 """
+
 from pathlib import Path
 
 try:
@@ -54,12 +55,16 @@ class CaimanSegmentationExtractor(SegmentationExtractor):
         SegmentationExtractor.__init__(self)
         self.file_path = file_path
         self._dataset_file = self._file_extractor_read()
+        self._roi_response_raw = self._raw_trace_extractor_read()
         self._roi_response_dff = self._trace_extractor_read("F_dff")
-        self._roi_response_neuropil = self._trace_extractor_read("C")
+        self._roi_response_denoised = self._trace_extractor_read("C")
+        self._roi_response_neuropil = self._trace_extractor_read("f")
         self._roi_response_deconvolved = self._trace_extractor_read("S")
-        self._image_correlation = self._summary_image_read()
+        self._image_correlation = self._correlation_image_read()
+        self._image_mean = self._summary_image_read()
         self._sampling_frequency = self._dataset_file["params"]["data"]["fr"][()]
         self._image_masks = self._image_mask_sparse_read()
+        self._background_image_masks = self._background_image_mask_read()
 
     def __del__(self):  # TODO: refactor segmentation extractors who use __del__ together into a base class
         """Close the h5py file when the object is deleted."""
@@ -93,6 +98,19 @@ class CaimanSegmentationExtractor(SegmentationExtractor):
         image_masks = np.reshape(image_mask_in, (*self.get_image_size(), -1), order="F")
         return image_masks
 
+    def _background_image_mask_read(self):
+        """Read the image masks from the h5py file.
+
+        Returns
+        -------
+        image_masks: numpy.ndarray
+            The image masks for each background components.
+        """
+        if self._dataset_file["estimates"].get("b"):
+            background_image_mask_in = self._dataset_file["estimates"]["b"]
+            background_image_masks = np.reshape(background_image_mask_in, (*self.get_image_size(), -1), order="F")
+            return background_image_masks
+
     def _trace_extractor_read(self, field):
         """Read the traces specified by the field from the estimates dataset of the h5py file.
 
@@ -111,10 +129,28 @@ class CaimanSegmentationExtractor(SegmentationExtractor):
         if field in self._dataset_file["estimates"]:
             return lazy_ops.DatasetView(self._dataset_file["estimates"][field]).lazy_transpose()
 
-    def _summary_image_read(self):
-        """Read the summary image (Cn) from the estimates dataset of the h5py file."""
+    def _raw_trace_extractor_read(self):
+        """Read the denoised trace and the residual trace from the h5py file and sum them to obtain the raw roi response trace.
+
+        Returns
+        -------
+        roi_response_raw: numpy.ndarray
+            The raw roi response trace.
+        """
+        roi_response_raw = self._dataset_file["estimates"]["C"][:] + self._dataset_file["estimates"]["YrA"][:]
+        return np.array(roi_response_raw.T)
+
+    def _correlation_image_read(self):
+        """Read correlation image Cn."""
         if self._dataset_file["estimates"].get("Cn"):
             return np.array(self._dataset_file["estimates"]["Cn"])
+
+    def _summary_image_read(self):
+        """Read summary image mean."""
+        if self._dataset_file["estimates"].get("b"):
+            FOV_shape = self._dataset_file["params"]["data"]["dims"][()]
+            b_sum = self._dataset_file["estimates"]["b"][:].sum(axis=1)
+            return np.array(b_sum).reshape(FOV_shape, order="F")
 
     def get_accepted_list(self):
         accepted = self._dataset_file["estimates"]["idx_components"]
@@ -176,8 +212,10 @@ class CaimanSegmentationExtractor(SegmentationExtractor):
             estimates = f.create_group("estimates")
             params = f.create_group("params")
             # adding to estimates:
+            if segmentation_object.get_traces(name="denoised") is not None:
+                estimates.create_dataset("C", data=segmentation_object.get_traces(name="denoised"))
             if segmentation_object.get_traces(name="neuropil") is not None:
-                estimates.create_dataset("C", data=segmentation_object.get_traces(name="neuropil"))
+                estimates.create_dataset("f", data=segmentation_object.get_traces(name="neuropil"))
             if segmentation_object.get_traces(name="dff") is not None:
                 estimates.create_dataset("F_dff", data=segmentation_object.get_traces(name="dff"))
             if segmentation_object.get_traces(name="deconvolved") is not None: