+++ src/roiextractors/extractors/inscopixextractors/inscopiximagingextractor.py
@@ -1,6 +1,4 @@
-"""
-Inscopix Imaging Extractor
-"""
+"""Inscopix Imaging Extractor."""
 
 import warnings
 from typing import Optional, Tuple
@@ -18,6 +16,8 @@ class InscopixImagingExtractor(ImagingExtractor):
 
     def __init__(self, file_path: PathType):
         """
+        Create an InscopixImagingExtractor instance from a single .isx file.
+
         Parameters
         ----------
         file_path : PathType