+++ src/roiextractors/multisegmentationextractor.py
@@ -5,6 +5,7 @@ Classes
 MultiSegmentationExtractor
     This class is used to combine multiple SegmentationExtractor objects by frames.
 """
+
 import numpy as np
 
 from .segmentationextractor import SegmentationExtractor