+++ tests/test_scanimagetiffimagingextractor.py
@@ -1,9 +1,26 @@
 import pytest
 from numpy.testing import assert_array_equal
-from ScanImageTiffReader import ScanImageTiffReader
-from roiextractors import ScanImageTiffSinglePlaneImagingExtractor, ScanImageTiffMultiPlaneImagingExtractor
 
+from roiextractors import (
+    ScanImageTiffSinglePlaneImagingExtractor,
+    ScanImageTiffMultiPlaneImagingExtractor,
+    ScanImageTiffSinglePlaneMultiFileImagingExtractor,
+    ScanImageTiffMultiPlaneMultiFileImagingExtractor,
+)
+from roiextractors.extractors.tiffimagingextractors.scanimagetiff_utils import (
+    extract_extra_metadata,
+    parse_metadata,
+    _get_scanimage_reader,
+)
 from .setup_paths import OPHYS_DATA_PATH
+
+import platform
+
+is_m_series_mac = platform.system() == "Darwin" and platform.machine() == "arm64"
+if (
+    is_m_series_mac
+):  # Remove this check once scanimage tiff reader is available on ARM -- see https://gitlab.com/vidriotech/scanimagetiffreader-python/-/issues/31
+    pytest.skip("ScanImageTiffReader does not support M-series Macs", allow_module_level=True)
 
 
 @pytest.fixture(scope="module")
@@ -45,8 +62,48 @@ def test_ScanImageTiffSinglePlaneImagingExtractor__init__invalid(file_path, chan
         ScanImageTiffSinglePlaneImagingExtractor(file_path=file_path, channel_name=channel_name, plane_name=plane_name)
 
 
+def test_ScanImageTiffSinglePlaneImagingExtractor__init__metadata_provided(file_path):
+    metadata = extract_extra_metadata(file_path)
+    parsed_metadata = parse_metadata(metadata)
+    extractor = ScanImageTiffSinglePlaneImagingExtractor(
+        file_path=file_path,
+        channel_name="Channel 1",
+        plane_name="0",
+        metadata=metadata,
+        parsed_metadata=parsed_metadata,
+    )
+    assert extractor.metadata == metadata
+    assert extractor.parsed_metadata == parsed_metadata
+
+
+def test_ScanImageTiffSinglePlaneImagingExtractor__init__invalid_metadata_provided(file_path):
+    metadata = {"invalid_key": "invalid_value"}
+    parsed_metadata = {"invalid_key": "invalid_value"}
+    with pytest.raises(KeyError):
+        ScanImageTiffSinglePlaneImagingExtractor(
+            file_path=file_path,
+            channel_name="Channel 1",
+            plane_name="0",
+            metadata=metadata,
+            parsed_metadata=parsed_metadata,
+        )
+
+
+def test_ScanImageTiffSinglePlaneImagingExtractor__init__parsed_metadata_not_provided(file_path):
+    metadata = extract_extra_metadata(file_path)
+    with pytest.raises(AssertionError):
+        ScanImageTiffSinglePlaneImagingExtractor(
+            file_path=file_path,
+            channel_name="Channel 1",
+            plane_name="0",
+            metadata=metadata,
+        )
+
+
 @pytest.mark.parametrize("frame_idxs", (0, [0, 1, 2], [0, 2, 5]))
 def test_get_frames(scan_image_tiff_single_plane_imaging_extractor, frame_idxs, expected_properties):
+    ScanImageTiffReader = _get_scanimage_reader()
+
     frames = scan_image_tiff_single_plane_imaging_extractor.get_frames(frame_idxs=frame_idxs)
     file_path = str(scan_image_tiff_single_plane_imaging_extractor.file_path)
     plane = scan_image_tiff_single_plane_imaging_extractor.plane
@@ -56,6 +113,7 @@ def test_get_frames(scan_image_tiff_single_plane_imaging_extractor, frame_idxs, 
     frames_per_slice = expected_properties["frames_per_slice"]
     if isinstance(frame_idxs, int):
         frame_idxs = [frame_idxs]
+
     raw_idxs = []
     for idx in frame_idxs:
         cycle = idx // frames_per_slice
@@ -67,6 +125,7 @@ def test_get_frames(scan_image_tiff_single_plane_imaging_extractor, frame_idxs, 
             + channel
         )
         raw_idxs.append(raw_idx)
+
     with ScanImageTiffReader(file_path) as io:
         assert_array_equal(frames, io.data()[raw_idxs])
 
@@ -79,6 +138,8 @@ def test_get_frames_invalid(scan_image_tiff_single_plane_imaging_extractor, fram
 
 @pytest.mark.parametrize("frame_idx", (1, 3, 5))
 def test_get_single_frame(scan_image_tiff_single_plane_imaging_extractor, expected_properties, frame_idx):
+    ScanImageTiffReader = _get_scanimage_reader()
+
     frame = scan_image_tiff_single_plane_imaging_extractor._get_single_frame(frame=frame_idx)
     file_path = str(scan_image_tiff_single_plane_imaging_extractor.file_path)
     plane = scan_image_tiff_single_plane_imaging_extractor.plane
@@ -111,6 +172,8 @@ def test_get_video(
     start_frame,
     end_frame,
 ):
+    ScanImageTiffReader = _get_scanimage_reader()
+
     video = scan_image_tiff_single_plane_imaging_extractor.get_video(start_frame=start_frame, end_frame=end_frame)
     if start_frame is None:
         start_frame = 0
@@ -122,6 +185,7 @@ def test_get_video(
     num_planes = expected_properties["num_planes"]
     num_channels = expected_properties["num_channels"]
     frames_per_slice = expected_properties["frames_per_slice"]
+
     raw_idxs = []
     for idx in range(start_frame, end_frame):
         cycle = idx // frames_per_slice
@@ -133,6 +197,7 @@ def test_get_video(
             + channel
         )
         raw_idxs.append(raw_idx)
+
     with ScanImageTiffReader(file_path) as io:
         assert_array_equal(video, io.data()[raw_idxs])
 
@@ -230,3 +295,87 @@ def test_ScanImageTiffMultiPlaneImagingExtractor__init__(file_path):
 def test_ScanImageTiffMultiPlaneImagingExtractor__init__invalid(file_path):
     with pytest.raises(ValueError):
         ScanImageTiffMultiPlaneImagingExtractor(file_path=file_path, channel_name="Invalid Channel")
+
+
+def test_ScanImageTiffMultiPlaneImagingExtractor__init__metadata_provided(file_path):
+    metadata = extract_extra_metadata(file_path)
+    parsed_metadata = parse_metadata(metadata)
+    extractor = ScanImageTiffMultiPlaneImagingExtractor(
+        file_path=file_path, metadata=metadata, parsed_metadata=parsed_metadata
+    )
+    assert extractor.metadata == metadata
+    assert extractor.parsed_metadata == parsed_metadata
+
+
+def test_ScanImageTiffMultiPlaneImagingExtractor__init__invalid_metadata_provided(file_path):
+    metadata = {"invalid_key": "invalid_value"}
+    parsed_metadata = {"invalid_key": "invalid_value"}
+    with pytest.raises(KeyError):
+        ScanImageTiffMultiPlaneImagingExtractor(file_path=file_path, metadata=metadata, parsed_metadata=parsed_metadata)
+
+
+@pytest.fixture(scope="module")
+def scanimage_folder_path():
+    return OPHYS_DATA_PATH / "imaging_datasets" / "ScanImage"
+
+
+@pytest.fixture(scope="module")
+def multifile_file_pattern():
+    return "scanimage_20240320_multifile_*.tif"
+
+
+@pytest.fixture(scope="module")
+def expected_file_names():
+    return [
+        "scanimage_20240320_multifile_00001.tif",
+        "scanimage_20240320_multifile_00002.tif",
+        "scanimage_20240320_multifile_00003.tif",
+    ]
+
+
+@pytest.mark.parametrize("extract_all_metadata", [True, False])
+def test_ScanImageTiffSinglePlaneMultiFileImagingExtractor__init__(
+    scanimage_folder_path, multifile_file_pattern, expected_file_names, extract_all_metadata
+):
+    extractor = ScanImageTiffSinglePlaneMultiFileImagingExtractor(
+        folder_path=scanimage_folder_path,
+        file_pattern=multifile_file_pattern,
+        channel_name="Channel 1",
+        plane_name="0",
+        extract_all_metadata=extract_all_metadata,
+    )
+    file_names = [imaging_extractor.file_path.name for imaging_extractor in extractor._imaging_extractors]
+    assert file_names == expected_file_names
+
+
+def test_ScanImageTiffSinglePlaneMultiFileImagingExtractor__init__invalid(scanimage_folder_path):
+    with pytest.raises(ValueError):
+        ScanImageTiffSinglePlaneMultiFileImagingExtractor(
+            folder_path=scanimage_folder_path,
+            file_pattern="invalid_pattern",
+            channel_name="Channel 1",
+            plane_name="0",
+        )
+
+
+@pytest.mark.parametrize("extract_all_metadata", [True, False])
+def test_ScanImageTiffMultiPlaneMultiFileImagingExtractor__init__(
+    scanimage_folder_path, multifile_file_pattern, expected_file_names, extract_all_metadata
+):
+    extractor = ScanImageTiffMultiPlaneMultiFileImagingExtractor(
+        folder_path=scanimage_folder_path,
+        file_pattern=multifile_file_pattern,
+        channel_name="Channel 1",
+        extract_all_metadata=extract_all_metadata,
+    )
+    file_names = [imaging_extractor.file_path.name for imaging_extractor in extractor._imaging_extractors]
+    assert file_names == expected_file_names
+
+
+def test_ScanImageTiffMultiPlaneMultiFileImagingExtractor__init__invalid(scanimage_folder_path):
+    with pytest.raises(ValueError):
+        ScanImageTiffMultiPlaneMultiFileImagingExtractor(
+            folder_path=scanimage_folder_path,
+            file_pattern="invalid_pattern",
+            channel_name="Channel 1",
+        )